pyproject.toml
[project]
name = "2025-google-capstone"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.18.0",
]
